refactor(components): move code-engine components to chat folder
