+++ frontend/src/app/chat/page.tsx
@@ -1,5 +1,5 @@
-import Home from './Home';
+import Chat from '@/components/chat';
 
 export default function Page() {
-  return <Home />;
+  return <Chat />;
 }
