+++ frontend/src/components/chat/chat-panel.tsx
