+++ frontend/src/components/chat/code-engine/file-structure.tsx
