+++ frontend/src/components/chat/project-modal.tsx
